+++ front/common/utils/track.js
@@ -1,11 +1,7 @@
 /* eslint no-console: 0 */
 
-import Comscore from '../modules/Trackers/Comscore';
 import Internal from '../modules/Trackers/Internal';
-import IVW3 from '../modules/Trackers/IVW3';
 import Krux from '../modules/Trackers/Krux';
-import Nielsen from '../modules/Trackers/Nielsen';
-import Quantserve from '../modules/Trackers/Quantserve';
 import UniversalAnalytics from '../modules/Trackers/UniversalAnalytics';
 
 /**
@@ -34,12 +30,8 @@ import UniversalAnalytics from '../modules/Trackers/UniversalAnalytics';
  */
 
 const trackers = {
-		Comscore,
 		Internal,
-		IVW3,
 		Krux,
-		Nielsen,
-		Quantserve,
 		UniversalAnalytics
 	},
 	/**
@@ -200,6 +192,15 @@ export function trackPageView(adsContext) {
 			instance.trackPageView(instance.usesAdsContext ? adsContext : context);
 		}
 	});
+
+	if (M.prop('initialPageView')) {
+		M.prop('initialPageView', false);
+	} else {
+		window.trackQuantservePageView();
+		window.trackNielsenPageView();
+		window.trackComscorePageView();
+		window.trackIVW3PageView();
+	}
 }
 
 /**