+++ front/main/app/components/infobox-builder-sidebar.js
@@ -9,18 +9,16 @@ export default Ember.Component.extend({
 			icon: 'row-block'
 		},
 		{
+			name: 'image',
+			icon: 'image-block'
+		},
+		{
 			name: 'title',
 			icon: 'title-block'
 		},
 		{
-			name: 'image',
-			icon: 'image-block'
+			name: 'section-header',
+			icon: 'section-block'
 		}
-		// this infobox item will be added as part of DAT-3711
-		// https://wikia-inc.atlassian.net/browse/DAT-3711
-		// {
-		//	name: 'section-header',
-		//	icon: 'section-block'
-		// }
 	]
 });