+++ front/main/app/mixins/infobox-builder-item.js
@@ -2,11 +2,16 @@ import Ember from 'ember';
 
 export default Ember.Mixin.create({
 	classNameBindings: ['active'],
-	active: Ember.computed('item', 'activeItem', function () {
-		return this.get('item') === this.get('activeItem');
+
+	active: Ember.computed('model', 'activeItem', function () {
+		return this.get('model') === this.get('activeItem');
 	}),
 
-	click() {
-		this.get('setEditItem')(this.get('item'));
+	mouseMove(event) {
+		this.get('onMouseEnter')(event.clientX, event.clientY);
+	},
+
+	mouseLeave() {
+		this.get('onMouseLeave')();
 	}
 });