+++ front/main/app/models/infobox-builder.js
@@ -9,7 +9,8 @@ const InfoboxBuilderModel = Ember.Object.extend({
 		this._itemIndex = {
 			row: 0,
 			image: 0,
-			title: 0
+			title: 0,
+			'section-header': 0
 		};
 		this.infoboxState = [];
 		this.itemInEditMode = null;
@@ -23,7 +24,6 @@ const InfoboxBuilderModel = Ember.Object.extend({
 	 */
 	addToState(object) {
 		this.get('infoboxState').pushObject(object);
-
 		return object;
 	},
 
@@ -39,18 +39,21 @@ const InfoboxBuilderModel = Ember.Object.extend({
 		let item = {};
 
 		switch (type) {
-		case 'title':
-			item = InfoboxBuilderModel.extendTitleData(this.createTitleItem(), elementData);
-			break;
-		case 'row':
-			item = InfoboxBuilderModel.extendRowData(this.createRowItem(), elementData);
-			break;
-		case 'image':
-			item = InfoboxBuilderModel.extendRowData(this.createImageItem(), elementData);
-			break;
-		default:
-			Ember.Logger.warn(`Unsupported infobox builder type encountered: '${type}'`);
-			break;
+			case 'title':
+				item = InfoboxBuilderModel.extendTitleData(this.createTitleItem(), elementData);
+				break;
+			case 'row':
+				item = InfoboxBuilderModel.extendRowData(this.createRowItem(), elementData);
+				break;
+			case 'image':
+				item = InfoboxBuilderModel.extendRowData(this.createImageItem(), elementData);
+				break;
+			case 'section-header':
+				item = InfoboxBuilderModel.extendHeaderData(this.createSectionHeaderItem(), elementData);
+				break;
+			default:
+				Ember.Logger.warn(`Unsupported infobox builder type encountered: '${type}'`);
+				break;
 		}
 
 		return this.addToState(item);
@@ -78,7 +81,8 @@ const InfoboxBuilderModel = Ember.Object.extend({
 				component: InfoboxBuilderModel.createComponentName(itemType)
 			},
 			source: `${itemType}${index}`,
-			type: itemType
+			type: itemType,
+			sourceFrozen: false
 		};
 	},
 
@@ -115,7 +119,7 @@ const InfoboxBuilderModel = Ember.Object.extend({
 	 */
 	createTitleItem() {
 		const itemType = 'title',
-			index = this.increaseItemIndex('title');
+			index = this.increaseItemIndex(itemType);
 
 		return {
 			data: {
@@ -126,6 +130,24 @@ const InfoboxBuilderModel = Ember.Object.extend({
 				component: InfoboxBuilderModel.createComponentName(itemType)
 			},
 			source: `${itemType}${index}`,
+			type: itemType
+		};
+	},
+
+	createSectionHeaderItem() {
+		const itemType = 'section-header',
+			index = this.increaseItemIndex(itemType);
+
+		return {
+			data: i18n.t('main.section-header-default', {
+				ns: 'infobox-builder',
+				index
+			}),
+			collapsible: false,
+			infoboxBuilderData: {
+				index,
+				component: InfoboxBuilderModel.createComponentName(itemType)
+			},
 			type: itemType
 		};
 	},
@@ -176,9 +198,23 @@ const InfoboxBuilderModel = Ember.Object.extend({
 
 		this.set(`infoboxState.${index}.data.label`, value);
 
-		if (value.trim().length) {
+		if (!item.sourceFrozen && value.trim().length) {
 			this.set(`infoboxState.${index}.source`, InfoboxBuilderModel.sanitizeCustomRowSource(value));
 		}
+	},
+
+	/**
+	 * @desc sets a new value of the data field
+	 * on the given section header element
+	 *
+	 * @param {Object} item
+	 * @param {Object} newValues
+	 * @returns {void}
+	 */
+	editSectionHeaderItem(item, newValues) {
+		const index = this.get('infoboxState').indexOf(item);
+
+		Object.keys(newValues).forEach((key) => this.set(`infoboxState.${index}.${key}`, newValues[key]));
 	},
 
 	/**
@@ -188,44 +224,16 @@ const InfoboxBuilderModel = Ember.Object.extend({
 	 */
 	removeItem(item) {
 		this.get('infoboxState').removeObject(item);
-		this.resetEditMode();
+		this.setEditItem(null);
 	},
 
 	/**
-	 * @desc moves item in infoboxState by given offset
-	 * @param {Number} offset
-	 * @param {Object} item
+	 * @desc updates infobox state order
+	 * @param {Ember.Array} newState
 	 * @returns {void}
 	 */
-	moveItem(offset, item) {
-		const position = this.get('infoboxState').indexOf(item);
-
-		if (this.isValidMove(position, offset)) {
-			this.get('infoboxState').removeAt(position);
-			this.get('infoboxState').insertAt(position + offset, item);
-		}
-	},
-
-	/**
-	 * @desc checks if move is valid based on item current position in the infoboxState and the move offset
-	 * @param {Number} position
-	 * @param {Number} offset
-	 * @returns {Boolean}
-	 */
-	isValidMove(position, offset) {
-		const lastItemIndex = this.get('infoboxState').length - 1,
-			newPosition = position + offset;
-
-		return position > 0 && offset < 0 && newPosition >= 0 ||
-			position < lastItemIndex && offset > 0 && newPosition <= lastItemIndex;
-	},
-
-	/**
-	 * @desc resets item in edit mode and its position to null
-	 * @returns {void}
-	 */
-	resetEditMode() {
-		this.set('itemInEditMode', null);
+	updateInfoboxStateOrder(newState) {
+		this.set('infoboxState', newState);
 	},
 
 	/**
@@ -330,16 +338,14 @@ InfoboxBuilderModel.reopenClass({
 	 */
 	extendRowData(item, itemData) {
 		if (itemData) {
+			const {data} = itemData,
+				// as data can be devoid of label value
+				{label} = data || {};
+
 			item.source = itemData.source || '';
-			item.data.label = '';
-
-			if (itemData.data) {
-				const {data: {label}} = itemData;
-
-				item.data.label = label || '';
-			}
+			item.data.label = label || '';
+			item.sourceFrozen = true;
 		}
-
 		return item;
 	},
 
@@ -355,14 +361,12 @@ InfoboxBuilderModel.reopenClass({
 	 */
 	extendTitleData(item, itemData) {
 		if (itemData) {
+			const {data} = itemData,
+				// as title can be devoid of default value
+				{defaultValue} = data || {};
+
 			item.source = itemData.source || '';
-			item.data.defaultValue = '';
-
-			if (itemData.data) {
-				const {data: {defaultValue}} = itemData;
-
-				item.data.defaultValue = defaultValue || '';
-			}
+			item.data.defaultValue = defaultValue || '';
 		}
 
 		return item;
@@ -371,7 +375,7 @@ InfoboxBuilderModel.reopenClass({
 	/**
 	 * @desc Overrides some properties of given Image object with additional
 	 * data, obtained from already existing template
-	 * TODO: use Object.assign() when we switch to Babel6
+	 * @todo use Object.assign() when we switch to Babel6
 	 * https://wikia-inc.atlassian.net/browse/DAT-3825
 	 *
 	 * @param {Object} item item to extend
@@ -388,6 +392,23 @@ InfoboxBuilderModel.reopenClass({
 
 				item.data.caption.source = captionSource || '';
 			}
+		}
+
+		return item;
+	},
+
+	/**
+	 * @desc Overrides some properties of given header object with additional
+	 * data, obtained from already existing template
+	 *
+	 * @param {Object} item item to extend
+	 * @param {Object} itemData additional data
+	 * @returns {Object}
+	 */
+	extendHeaderData(item, itemData) {
+		if (itemData) {
+			item.data = itemData.data || '';
+			item.collapsible = itemData.collapsible || false;
 		}
 
 		return item;