+++ front/main/app/models/mediawiki/category.js
@@ -1,6 +1,0 @@
-import Ember from 'ember';
-
-const CategoryModel = Ember.Object.extend({
-});
-
-export default CategoryModel;