+++ server/app/facets/mediaWikiPage.js
@@ -7,6 +7,7 @@ import * as Utils from '../lib/Utils';
 import getStatusCode from './operations/getStatusCode';
 import localSettings from '../../config/localSettings';
 import prepareArticleData from './operations/prepareArticleData';
+import prepareCategoryData from './operations/prepareCategoryData';
 import prepareMainPageData from './operations/prepareMainPageData';
 import prepareMediaWikiData from './operations/prepareMediaWikiData';
 import deepExtend from 'deep-extend';
@@ -55,7 +56,7 @@ function redirectToMainPage(reply, mediaWikiPageHelper) {
 }
 
 /**
- * Handles article response from API
+ * Handles getPage response from API
  *
  * @param {Hapi.Request} request
  * @param {Hapi.Response} reply
@@ -65,32 +66,55 @@ function redirectToMainPage(reply, mediaWikiPageHelper) {
  * @returns {void}
  */
 function handleResponse(request, reply, data, allowCache = true, code = 200) {
-	const ns = data.page.data.ns;
+	const i18n = request.server.methods.i18n.getInstance();
 
-	let result, response;
+	let result = {},
+		pageData = {},
+		viewName = 'wiki-page',
+		response,
+		ns;
 
-	switch (ns) {
-	case MediaWikiNamespace.MAIN:
-		result = prepareArticleData(request, data);
-
-		// mainPageData is set only on curated main pages - only then we should do some special preparation for data
-		if (data.page.data && data.page.data.isMainPage && data.page.data.mainPageData) {
-			result = deepExtend(result, prepareMainPageData(data));
-			delete result.adsContext;
-		}
-		break;
-
-	default:
-		Logger.info(`Unsupported namespace: ${ns}`);
-		result = prepareMediaWikiData(request, data);
+	if (data.page && data.page.data) {
+		pageData = data.page.data;
+		ns = pageData.ns;
+		result.mediaWikiNamespace = ns;
 	}
 
-	result.mediaWikiNamespace = ns;
+	switch (ns) {
+		case MediaWikiNamespace.MAIN:
+			viewName = 'article';
+			result = deepExtend(result, prepareArticleData(request, data));
+
+			break;
+
+		case MediaWikiNamespace.CATEGORY:
+			if (pageData.article && pageData.details) {
+				viewName = 'article';
+				result = deepExtend(result, prepareArticleData(request, data));
+			}
+
+			result = deepExtend(result, prepareCategoryData(request, data));
+			// Hide TOC on category pages
+			result.hasToC = false;
+			result.subtitle = i18n.t('app.category-page-subtitle');
+			break;
+
+		default:
+			Logger.info(`Unsupported namespace: ${ns}`);
+			result = prepareMediaWikiData(request, data);
+	}
+
+	// mainPageData is set only on curated main pages - only then we should do some special preparation for data
+	if (pageData.isMainPage && pageData.mainPageData) {
+		result = deepExtend(result, prepareMainPageData(data));
+		result.hasToC = false;
+		delete result.adsContext;
+	}
 
 	// @todo XW-596 we shouldn't rely on side effects of this function
 	Tracking.handleResponse(result, request);
 
-	response = reply.view('article', result);
+	response = reply.view(viewName, result);
 	response.code(code);
 	response.type('text/html; charset=utf-8');
 
@@ -102,7 +126,7 @@ function handleResponse(request, reply, data, allowCache = true, code = 200) {
 }
 
 /**
- * Gets wiki variables and article, handles errors on both promises
+ * Gets wiki variables and wiki page, handles errors on both promises
  *
  * @param {Hapi.Request} request
  * @param {Hapi.Response} reply
@@ -135,15 +159,15 @@ function getMediaWikiPage(request, reply, mediaWikiPageHelper, allowCache) {
 		 */
 		.catch(PageRequestError, (error) => {
 			const data = error.data,
-				errorCode = getStatusCode(data.article, 500);
+				errorCode = getStatusCode(data.page, 500);
 
-			Logger.error(data.article.exception, 'MediaWikiPage error');
+			Logger.error(data.page.exception, 'MediaWikiPage error');
 
 			// It's possible that the article promise is rejected but we still want to redirect to canonical host
 			Utils.redirectToCanonicalHostIfNeeded(localSettings, request, reply, data.wikiVariables);
 
 			// Clean up exception to not put its details in HTML response
-			delete data.article.exception.details;
+			delete data.page.exception.details;
 
 			handleResponse(request, reply, data, allowCache, errorCode);
 		})